requirements.txt
fastapi
python-dotenv
python-jose
sse-starlette
torch>=2.0
transformers==4.31.0
transformers_stream_generator==0.0.4
uvicorn
sentencepiece
accelerate>=0.20.3

# for Stable Diffusion
diffusers
invisible_watermark
safetensors

python-multipart